estatex/backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
redis==5.0.3
django-redis==5.4.0
django-environ==0.11.2
gunicorn==21.2.0
pytest==8.1.1
pytest-django==4.8.0
